♻️ : zustand로 newComment, post는(일단 zustand 안씀)
